+++ src/parser/NodeLine.java
@@ -16,7 +16,7 @@ public class NodeLine extends ResponsibleChainNode{
      * 绘制线段
      */
     @Override
-    protected void exec(GeomCompos geom, ArrayList<Integer> args) {
+    protected void exec(GeomCompos geom, ArrayList<Float> args) {
         Points p1 = new Points(args.get(0),args.get(1));
         Points p2 = new Points(args.get(2),args.get(3));
         Line line = new Line(p1,p2);
